+++ internal/models/messages.go
@@ -5,7 +5,7 @@ import "time"
 type Message struct {
 	ID        int64     `db:"id"`
 	ChatID    int64     `db:"chat_id"`
-	Number    int       `db:"number"`
+	Number    int64     `db:"number"`
 	Body      string    `db:"body"`
 	CreatedAt time.Time `db:"created_at"`
 	UpdatedAt time.Time `db:"updated_at"`
